Moving 'createActionHandler' and 'triggerBehaviors' into new /services/behaviors/helpers file
